Address CodeRabbit nitpicks on the indicator reshape (CF-453)
Internal EventSynchronizationResult becomes a record with plain List.copyOf
copies; same-series validation is shared via the support helper; the
benchmark helper drops its unused shift parameter; the predicate-boundary
test moves to the support test; the two brute-force oracle sweeps merge into
one exhaustive owner with the worst-offset assertion.
